+++ CHANGELOG.md
@@ -8,6 +8,12 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Fixed
+
+- Handle video not found  error when task to transcode video
+  by increasing retries in celery's task
+
+
 ## [5.12.4] - 2026-07-20
 
 ### Fixed
@@ -283,7 +289,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Meta description and meta title on the website from the API (#2516)
 - Retrieve BBB learning analytics and send them through API
-- Classroom attendance analytics (#2499) 
+- Classroom attendance analytics (#2499)
 - Add a language picker for the invite link on the website (#2504)
 
 ### Changed